requirements.txt
absl-py==0.8.0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.4
astor==0.8.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
bcrypt==3.1.4
cachetools==4.2.0
certifi==2019.9.11
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.1.0
django-crispy-forms==1.7.2
django-dbbackup==1.0
django-excel==0.0.10
django-payu==0.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.0.12
flatbuffers==1.12
gast==0.3.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
hiredis==1.1.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==3.3.0
incremental==17.5.0
jdcal==1.4.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lml==0.0.9
Markdown==3.1.1
msgpack==1.0.0
mysqlclient==1.4.6
numpy==1.17.2
oauthlib==3.1.0
olefile==0.44
openpyxl==3.0.2
opt-einsum==3.0.1
Pillow==6.1.0
pipenv==2018.10.13
protobuf==3.9.1
psycopg2==2.8.3
psycopg2-binary==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel==0.6.2
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pypiwin32==223
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
pywin32==300
reportlab==3.5.26
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.6
scikit-learn==0.24.2
scipy==1.6.3
service-identity==18.1.0
six==1.12.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tb-nightly==1.15.0a20190806
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.0.0rc1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
texttable==1.6.2
tf-estimator-nightly==1.14.0.dev2019080601
threadpoolctl==2.1.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.3
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.5
Werkzeug==0.15.6
wfastcgi==3.0.0
wrapt==1.11.2
xlrd==1.1.0
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.1.0
